Caching - Don't return dead entries from hashtable
Copied from - cfa32a9e09664ca3f56d13074cb9b52c4226b81f

When the refcount of an entry goes to zero, it's removed from the LRU
under the lane lock, and then the lane lock is dropped, and the entry is
cleaned up.  This includes removing it from the hash table.

However, there is a window during which cache_inode_find_by_name() or
cache_inode_new_entry() can find it from the hash table.

Fix the hash table to not return an entry if it's refcount is zero.

Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>