Do not collect latency stats for client
Statistics counting per client/export, currently involves,
operation and latency statistics for read, write commands.
While collecting and extracting stats per client, latency value has
no meaning. Actually it is not the latency from client perspective,
rather it is server perspective.
The per client stats would be helpful in getting insights of the
workload exerted by the client. For this information on number of
operations and kind of operations is helpful.
Not collecting latency information while gathering stats for a client.

Change-Id: I547b613fc94599d2854506283ff85febf30a5d7b
Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>